Sample Data & Tests
To connect our databases, we used employeeId and companyId as a primary foreign key from the employees table. This acts as a parent table to assigned_training, performance_review, and pto_request. After inserting various kinds of correct samples, we inserted samples where the employeeId and companyId were not in the parent table. This concludes that our foreign keys work as intended.